refactor k8s resource tools with factory and registry pattern for better structure
