+++ pkg/k8s/resourcetypes/helpers.go
@@ -1,0 +1,135 @@
+package resourcetypes
+
+import (
+	"fmt"
+
+	"github.com/mark3labs/mcp-go/mcp"
+)
+
+// RequiredParam is a helper function that can be used to fetch a required parameter from the request.
+// It checks if the parameter is present, of the expected type, and not empty (non-zero value).
+func RequiredParam[T comparable](r mcp.CallToolRequest, p string) (T, error) {
+	var zero T
+
+	// Check if the parameter is present in the request
+	if _, ok := r.Params.Arguments[p]; !ok {
+		return zero, fmt.Errorf("missing required parameter: %s", p)
+	}
+
+	// Check if the parameter is of the expected type
+	if _, ok := r.Params.Arguments[p].(T); !ok {
+		return zero, fmt.Errorf("parameter %s is not of type %T", p, zero)
+	}
+
+	if r.Params.Arguments[p].(T) == zero {
+		return zero, fmt.Errorf("missing required parameter: %s", p)
+	}
+
+	return r.Params.Arguments[p].(T), nil
+}
+
+// OptionalParam is a helper function that can be used to fetch an optional parameter from the request.
+// It checks if the parameter is present and of the expected type.
+func OptionalParam[T any](r mcp.CallToolRequest, p string) (T, error) {
+	var zero T
+
+	// Check if the parameter is present in the request
+	if _, ok := r.Params.Arguments[p]; !ok {
+		return zero, nil
+	}
+
+	// Check if the parameter is of the expected type
+	if _, ok := r.Params.Arguments[p].(T); !ok {
+		return zero, fmt.Errorf("parameter %s is not of type %T, is %T", p, zero, r.Params.Arguments[p])
+	}
+
+	return r.Params.Arguments[p].(T), nil
+}
+
+// OptionalIntParam is a helper function that can be used to fetch an optional integer parameter from the request.
+func OptionalIntParam(r mcp.CallToolRequest, p string) (int, error) {
+	v, err := OptionalParam[float64](r, p)
+	if err != nil {
+		return 0, err
+	}
+	return int(v), nil
+}
+
+// OptionalIntParamWithDefault is a helper function that can be used to fetch an optional integer parameter from the request
+// with a default value if the parameter is not present or zero.
+func OptionalIntParamWithDefault(r mcp.CallToolRequest, p string, d int) (int, error) {
+	v, err := OptionalIntParam(r, p)
+	if err != nil {
+		return 0, err
+	}
+	if v == 0 {
+		return d, nil
+	}
+	return v, nil
+}
+
+// OptionalStringArrayParam is a helper function that can be used to fetch an optional string array parameter from the request.
+func OptionalStringArrayParam(r mcp.CallToolRequest, p string) ([]string, error) {
+	// Check if the parameter is present in the request
+	if _, ok := r.Params.Arguments[p]; !ok {
+		return []string{}, nil
+	}
+
+	switch v := r.Params.Arguments[p].(type) {
+	case nil:
+		return []string{}, nil
+	case []string:
+		return v, nil
+	case []any:
+		strSlice := make([]string, len(v))
+		for i, v := range v {
+			s, ok := v.(string)
+			if !ok {
+				return []string{}, fmt.Errorf("parameter %s is not of type string, is %T", p, v)
+			}
+			strSlice[i] = s
+		}
+		return strSlice, nil
+	default:
+		return []string{}, fmt.Errorf("parameter %s could not be coerced to []string, is %T", p, r.Params.Arguments[p])
+	}
+}
+
+// WithPagination returns a ToolOption that adds "page" and "perPage" parameters to the tool.
+func WithPagination() mcp.ToolOption {
+	return func(tool *mcp.Tool) {
+		mcp.WithNumber("page",
+			mcp.Description("Page number for pagination (min 1)"),
+			mcp.Min(1),
+		)(tool)
+
+		mcp.WithNumber("perPage",
+			mcp.Description("Results per page for pagination (min 1, max 100)"),
+			mcp.Min(1),
+			mcp.Max(100),
+		)(tool)
+	}
+}
+
+// PaginationParams holds pagination parameters.
+type PaginationParams struct {
+	Page    int
+	PerPage int
+}
+
+// OptionalPaginationParams returns the "page" and "perPage" parameters from the request,
+// or their default values if not present.
+func OptionalPaginationParams(r mcp.CallToolRequest) (PaginationParams, error) {
+	page, err := OptionalIntParamWithDefault(r, "page", 1)
+	if err != nil {
+		return PaginationParams{}, err
+	}
+	perPage, err := OptionalIntParamWithDefault(r, "perPage", 30)
+	if err != nil {
+		return PaginationParams{}, err
+	}
+	return PaginationParams{
+		Page:    page,
+		PerPage: perPage,
+	}, nil
+}
